refactor projects by move all mainsite codes to root src folder
