+++ src/components/ImageWithTooltip.tsx
@@ -4,7 +4,7 @@ import type { FC, RefObject } from "react";
 import { useCallback, useContext, useRef, useState } from "react";
 import Skeleton from "react-loading-skeleton";
 import questionMarkSvg from "../../assets/question-mark-v2.svg";
-import { FeatureFlagsContext } from "../../feature-flags";
+import { FeatureFlagsContext } from "../contexts/FeatureFlagContext";
 
 type ImageWithTooltipProps = {
   className?: HTMLImageElement["className"];
