+++ src/components/Nerd.tsx
