+++ src/constants/duration.ts
