+++ src/constants/static-ether-data.ts
