+++ src/dashboards/Dashboards.tsx
@@ -5,16 +5,19 @@ import type { FC, MouseEvent } from "react";
 import { useCallback, useContext, useEffect, useRef, useState } from "react";
 import { SkeletonTheme } from "react-loading-skeleton";
 import "react-loading-skeleton/dist/skeleton.css";
-import colors from "../colors";
+import colors from "../constants/colors";
 import BasicErrorBoundary from "../components/BasicErrorBoundary";
 import HeaderGlow from "../components/HeaderGlow";
 import MainTitle from "../components/MainTitle";
-import { WEI_PER_ETH } from "../eth-units";
-import { FeatureFlagsContext, useFeatureFlags } from "../feature-flags";
-import { formatZeroDecimals } from "../format";
+import { WEI_PER_ETH } from "../constants/eth-units";
+import {
+  FeatureFlagsContext,
+  useFeatureFlags,
+} from "../contexts/FeatureFlagContext";
+import { formatZeroDecimals } from "../utils/format";
 import ContactSection from "./sections/ContactSection";
-import type { TimeFrame, getNextTimeFrame } from "../time-frames";
-import FaqBlock from "../mainsite/components/Faq";
+import type { TimeFrame, getNextTimeFrame } from "../utils/time-frames";
+import FaqBlock from "../components/Faq";
 import BurnDashboard from "./sections/BurnDashboard";
 import GasSection from "./sections/GasSection";
 import BlobGasSection from "./sections/BlobGasSection";
