+++ src/hooks/use-now.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { O } from "../fp";
+import { O } from "../utils/fp";
 
 export const useNow = () => {
   const [now, setNow] = useState<O.Option<Date>>(O.none);
